requirements.txt
sqlalchemy
sqlalchemy-redshift
pandas
pytest
psycopg2-binary
boto3
cryptography
sshtunnel
pymongo